\name{business-methods}
\docType{methods}
\alias{business-methods}
\alias{business,GCPM-method}
\alias{business}
\title{Counterparty Business Line}
\description{
Get the business information for each counterparty defined in the portfolio.
}
\usage{
  business(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{factor of length equal to number of portfolio positions}
\seealso{\code{\link{portfolio.pois}}}
\keyword{methods}
