\name{CDF-methods}
\docType{methods}
\alias{CDF-methods}
\alias{CDF}
\alias{CDF,GCPM-method}
\title{Cumulative Distribution Function of Portfolio Loss}
\description{
Get the CDF of the portfolio loss, available after execution of \code{analyze}. 
}
\usage{
  CDF(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric vector}
\seealso{\code{\link{analyze}}}
\keyword{methods}
