\name{country-methods}
\docType{methods}
\alias{country-methods}
\alias{country}
\alias{country,GCPM-method}
\title{Country Information}
\description{
Get the country information of each counterparty defined in the portfolio.}
\usage{
  country(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{factor of length equal to number of portfolio positions}
\seealso{\code{\link{portfolio.pois}}}
\keyword{methods}

