\name{EAD-methods}
\docType{methods}
\alias{EAD-methods}
\alias{EAD}
\alias{EAD,GCPM-method}
\title{Exposure at Default}
\description{
Get the counterparties' exposure at default defined in the portfolio data. 
}
\usage{
  EAD(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length equal to the number of counterparties}
\seealso{\code{\link{portfolio.pois}}}
\keyword{methods}
