\name{LGD-methods}
\docType{methods}
\alias{LGD-methods}
\alias{LGD}
\alias{LGD,GCPM-method}
\title{Loss Given Default}
\description{
Get the values of LGD, defined within the portfolio}
\usage{
LGD(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric vector of length equal to number of counterparties}
\seealso{\code{\link{portfolio.pois}}}
\keyword{methods}
