\name{link.function-methods}
\docType{methods}
\alias{link.function-methods}
\alias{link.function}
\alias{link.function,GCPM-method}
\title{Model Link Function}
\description{
Get the models link function (see \code{\link{init}})}
\usage{
link.function(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{character value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}
