\name{loss-methods}
\docType{methods}
\alias{loss-methods}
\alias{loss}
\alias{loss,GCPM-method}
\title{Loss Levels}
\description{
Get the loss levels of the portfolio loss distribution.
}
\usage{
loss(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric vector}
\keyword{methods}