\name{loss.thr-methods}
\docType{methods}
\alias{loss.thr-methods}
\alias{loss.thr}
\alias{loss.thr,GCPM-method}
\title{Threshold of Saved Portfolio Loss}
\description{
Get the value of \code{loss.thr} (see \code{\link{init}})
}
\usage{
loss.thr(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}
