\name{loss.unit-methods}
\docType{methods}
\alias{loss.unit-methods}
\alias{loss.unit}
\alias{loss.unit,GCPM-method}
\title{Loss Unit}
\description{
Get the loss unit used for potential loss discretization of the model}
\usage{
loss.unit(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}
