\name{model.type-methods}
\docType{methods}
\alias{model.type-methods}
\alias{model.type}
\alias{model.type,GCPM-method}
\title{Model Type}
\description{
Get the value of \code{model.type} (see \code{\link{init}})
}
\usage{
model.type(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{character value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}
