\name{N-methods}
\docType{methods}
\alias{N-methods}
\alias{N}
\alias{N,GCPM-method}
\title{Number of Simulations}
\description{
Get the value of \code{N} (number of simulations, see \code{\link{init}})
}
\usage{
N(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}
