\name{name-methods}
\docType{methods}
\alias{name-methods}
\alias{name}
\alias{name,GCPM-method}
\title{Counterparty Names}
\description{
Get the value of \code{name}, i.e. the counterparties' names, defined in the 
portfolio (see \code{\link{analyze}})
}
\usage{
name(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{character value of length equal to number of counterparties}
\seealso{\code{\link{portfolio.pois}}}
\keyword{methods}