\name{NC-methods}
\docType{methods}
\alias{NC-methods}
\alias{NC}
\alias{NC,GCPM-method}
\title{Number of Counterparties}
\description{
Get the value of \code{NC}, representing the number of counterparties within 
the portfolio (see \code{\link{analyze}})
}
\usage{
NC(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\seealso{\code{\link{analyze}}}
\keyword{methods}