\name{NR-methods}
\docType{methods}
\alias{NR-methods}
\alias{NR}
\alias{NR,GCPM-method}
\title{Counterparty IDs}
\description{
Get the value of \code{NR}, the counterparties' identification numbers within the 
portfolio (see \code{\link{analyze}})
}
\usage{
NR(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length equal to number of counterparties}
\seealso{\code{\link{portfolio.pois}}}
\keyword{methods}