\name{NS-methods}
\docType{methods}
\alias{NS-methods}
\alias{NS}
\alias{NS,GCPM-method}
\title{Number of Sectors}
\description{
Get the value of \code{NS}, the number of sectors within the model 
(see \code{\link{init}})
}
\usage{
NS(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}
