\name{PDF-methods}
\docType{methods}
\alias{PDF-methods}
\alias{PDF}
\alias{PDF,GCPM-method}
\title{Probability Density Function}
\description{
Get the value of \code{PDF}, representing the pdf of the estimated portfolio 
loss distribution.
}
\usage{
PDF(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric vector}
\keyword{methods}