\name{portfolios}
\alias{portfolios}
\docType{data}
\title{
Example Portfolios for GCPM Package}
\description{
The workspace contain the example portfolio (with Poisson default mode) in the structure needed by the \code{\link{analyze}} function as well as a pooled version. }
\usage{data("portfolios")}
\format{Two data frames containing the portfolios.}
\keyword{datasets}
\seealso{
\code{\link{portfolio.pois}}, \code{\link{portfolio.pool}}, \code{\link{analyze}}
}
