\name{random.numbers-methods}
\docType{methods}
\alias{random.numbers-methods}
\alias{random.numbers}
\alias{random.numbers,GCPM-method}
\title{Sector Drawings}
\description{
Get the content of \code{random.numbers}, representing the sector drawings 
(see \code{\link{init}})
}
\usage{
random.numbers(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric matrix}
\seealso{\code{\link{init}}}
\keyword{methods}