\name{SD-methods}
\docType{methods}
\alias{SD-methods}
\alias{SD}
\alias{SD,GCPM-method}
\title{Standard Deviation (Loss Distribution)}
\description{
Get the value of \code{SD}, the portfolio standard deviation derived from the 
loss distribution.
}
\usage{
SD(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\keyword{methods}