\name{SD.analyt-methods}
\docType{methods}
\alias{SD.analyt-methods}
\alias{SD.analyt}
\alias{SD.analyt,GCPM-method}
\title{Standard Deviation (from Portfolio Data)}
\description{
Get the value of \code{SD.analyt}, the portfolio standard deviation derived 
from the portfolio data (see \code{\link{GCPM-class}}). This value is only 
available in case of an analytical model. 
}
\usage{
SD.analyt(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\keyword{methods}