\name{SD.div-methods}
\docType{methods}
\alias{SD.div-methods}
\alias{SD.div}
\alias{SD.div,GCPM-method}
\title{Diversifiable Risk (Standard Deviation)}
\description{
Get the value of \code{SD.div}, the diversifiable part of portfolio standard
deviation (see \code{\link{GCPM-class}})
}
\usage{
SD.div(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\keyword{methods}