\name{SD.syst-methods}
\docType{methods}
\alias{SD.syst-methods}
\alias{SD.syst}
\alias{SD.syst,GCPM-method}
\title{Systemic Risk (Standard Deviation)}
\description{
Get the value of \code{SD.syst}, the non-diversifiable part of portfolio
standard deviation.
}
\usage{
SD.syst(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\keyword{methods}
