\name{sec.var-methods}
\docType{methods}
\alias{sec.var-methods}
\alias{sec.var}
\alias{sec.var,GCPM-method}
\title{Sector Variances}
\description{
Get the value of \code{sec.var}, the sector variances in case of an analytical CreditRisk+ 
like model (see \code{\link{init}})
}
\usage{
sec.var(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length equal to number of sectors}
\seealso{\code{\link{init}}}
\keyword{methods}