\name{sector.names-methods}
\docType{methods}
\alias{sector.names-methods}
\alias{sector.names}
\alias{sector.names,GCPM-method}
\title{Sector Names}
\description{
Get the value of \code{sector.names}, the sector names (see \code{\link{init}})
}
\usage{
sector.names(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{factor of length equal to number of sectors}
\seealso{\code{\link{init}}}
\keyword{methods}