\name{seed-methods}
\docType{methods}
\alias{seed-methods}
\alias{seed}
\alias{seed,GCPM-method}
\title{Random Number Seed}
\description{
Get the value of \code{seed} (see \code{\link{init}})
}
\usage{
seed(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric value of length 1}
\seealso{\code{\link{init}}}
\keyword{methods}