\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,ANY-method}
\alias{summary,GCPM-method}
\title{Model summary}
\description{
Create a Summary List with Model Parameters.}
\usage{
summary(object,...)
}
\arguments{
  \item{object}{Object of class \code{GCPM}}
  \item{...}{No further arguments}
}
\value{list}
\keyword{methods}

