\name{W-methods}
\docType{methods}
\alias{W-methods}
\alias{W}
\alias{W,GCPM-method}
\title{Sector Weights}
\description{
Get the value of \code{W}, the matrix of counterparties' sector weights defined 
within the portfolio (see \code{\link{analyze}})
}
\usage{
W(this)
}
\arguments{
  \item{this}{Object of class \code{GCPM}}
}
\value{numeric matrix}
\seealso{\code{\link{init}}}
\keyword{methods}
